veil sdk feature: taskpool (vtl1)
[Feature]

Provides a taskpool to enclaves that can queue tasks for futures with
exception support.

The vtl1 taskpool (veil::vtl1::taskpool) queues tasks and returns a future.
This is like std::async, but allows the user to specify how many threads
are in the pool to limit the total number of concurrent running tasks.

[Security & review note]

Please review and comment on all parts of the architecture and security
model, with the exception of enclave ABI parameter validation.

i.e. Call-in and callback data/buffer memcpying, bounds-checks + TOCTOU
safety is missing and waiting for the .edl tooling codegen work
implemented & consumed.

[Mechanics]

1. When a user calls queue_task(), the task is stored in an object_table
in the taskpool, in exchange for a task 'id'.

2. Then taskpool calls out to vtl0 with the task 'id' where it is
queued (in a std::dequeue). A store of real backing threads in
vtl0 are parked and waiting to try to schedule 'id's from the queue.

3. When an 'id' is scheduled by a newly-available thread, the thread
calls back into vtl1 with the task 'id', which can be fetched out
of the object_table and finally run.

Exception propagation is supported, which catches on the task
thread and rethrows on the thread calling future.get().

  See: taskpool.vtl1.h for more details

[Types]

- veil::vtl1::taskpool
- veil::vtl0::implementation::taskpool_backing_threads (free-floating thread pool)
- veil::vtl1::future  (std::future support missing)
- veil::vtl1::promise (std::promise support missing)
- veil::vtl1::keepalive_mechanism to keep the taskpool alive while child
    tasks are running.

[Testing]

Includes debug_print tests in my_exports.cpp sample code
for now (TAEF infra not yet online)

fix