Update vcpkg install in cpp_ci.yml (#155)
Got a notification that after checking in the cpp ci changes and the new
`optional<T>` changes to edl parser that the ci build of the edl codegen
failed. Good news is I got a failure email immediately which is nice 😊.

Issue that caused the failure was that the flatbuffer compiler wasn't
being built in the right place. So this path:
https://github.com/microsoft/VbsEnclaveTooling/blob/18e3eab05b669d40ee9e032b1a8703b87b0d102a/src/ToolingExecutable/edlcodegen.targets#L8
didn't exist in the pipeline.

This likely slipped through because even though I checked in the
`optional<T>` PR first, I should have merged my cpp ci branch with main
to get those changes and re-run the pipeline for my cpp ci branch. That
said, we should be good now since this PR branch is based on the most
up-to-date main that contains both changes and we're all green.