Update vpack version (#168)
### Why is this change needed?
I was doing some testing to confirm vpack creation works when I saw that
creation fails due to:
`##[error]ERROR: Cannot parse argument '251215.1' for option '/Patch' as
expected type 'System.Nullable`1[System.UInt64]'.
##[error]ERROR: `

This is because vpacks use semver 2.0 which specifies that the version
should be in the form `Major.Minor.Patch` with prerelease versions being
in the form `Major.Minor.Patch-AnyText.Number` E.g
`Major.Minor.Patch-prerelease.1`. I was using our nuget build number as
the version for the vpack which is in the form
`Major.Minor.Date.VersionCounter` E.g `0.2.251215.1`.

As you can see from the error above vpack assumes the patch version is a
`uint64` so the `.1` after the date caused it to fail.

To fix this, the patch version we will provide to vpack `will not` have
a dot between the date and the version counter like our nuget version.