Update pipeline templates and sdk project for OS consumption (#171)
### Why is this Change needed?
To allow for the codegen tool and the sdk to be used within the Windows
OS

### What changed?
- Updated the one branch official yml template to create an optional
vpack at the end of the build
- Split out the nuget packaging part of the build job templates into
their own optional job. One for the codegen package and one for the sdk
package. Now you can optionally choose to build a nuget package in the
ADO pipeline UI.
- Updated the codegen and sdk pipeline jobs to build x64/arm64 release
and debug configurations in parallel instead of sequentially.
- The SDK build job now uses the Windows Native compiler package when
the "Create VPack" option is selected in the ADO pipeline UI. This is
necessary for the sdk static libs to be consumed within the OS.
- Updated veil project vcxproj files to only use the Win SDK nuget
headers instead of using all the Win SDK headers + .props and .targets
files when being built in the Azure pipeline when VPack creation is
enabled. This prevents conflicts with the .props and .targets from the
Windows native compiler package.
- Removed .pdb additions to the nuget package as we will publish these
another way.
- fixed issue with flatc.exe not being compiled with control flow guard
by changing the host triplet to use our static cfg triplet by default.

### How was this tested?
- Although I can't officially test vpack creation as it requires this PR
to be approved first. I was able to confirm the builds succeed and nuget
packages are built with the correct data. [CodeGen
Result](https://microsoft.visualstudio.com/Dart/_build/results?buildId=136111056&view=results),
[SDK
Result](https://microsoft.visualstudio.com/Dart/_build/results?buildId=136111081&view=results)
- Once  this is checked in I can attempt to create the vpack officially.