Update sdk and samples with new codegen (#173)
### Why is this change needed?
We published a new codegen nuget package so the SDK + sample apps needs
to be updated to use it.

### What changed?
- Updated the packages.config and vcxproj files in sdk and samples to
use codegen `v0.2.251223.1`
- The sample enclave project failed to build for me due to it not
finding the `pch.h` file so I update it so it can find it by adding
`$(ProjectDir)` to the additional includes location and then changing
`"..\pch.h"` to `<pch.h>`
- Removed unneeded `VBS_Enclave_Dependencies` block in sample enclave's
vcxproj file. We have one at the bottom of the file that takes
precedence over it, so this one was never used. See it here:
https://github.com/microsoft/VbsEnclaveTooling/blob/c0b1387eaae2aeec2c25f0f62a43206dcab553af/SampleApps/SampleApps/SampleEnclave/SampleEnclave.vcxproj#L132
- Updated hello world doc based on issue #172 

### How was this tested?
- confirmed samples still build  and run correctly